updated changes

+++ src/screens/payments/BankAccounts.js
@@ -9,7 +9,8 @@ import {
   TouchableWithoutFeedback,
   Dimensions,
   Modal,
-  Pressable
+  Pressable,
+  Platform
 } from 'react-native';
 import MessageModal from '../../components/modals/MessageModal';
 import { useSelector } from 'react-redux';
@@ -25,9 +26,6 @@ import * as Keychain from 'react-native-keychain';
 import PoppinsText from '../../components/electrons/customFonts/PoppinsText';
 import { useDeleteBankMutation } from '../../apiServices/bankAccount.js/DeleteBankAccount';
 import { useIsFocused } from '@react-navigation/native';
-import { isAction } from '@reduxjs/toolkit';
-import { useUpdateStatusBankAccountMutation } from '../../apiServices/bankAccount.js/AddBankAccount';
-import FastImage from 'react-native-fast-image';
 
 
 const BankAccounts = ({ navigation, route }) => {
@@ -37,12 +35,7 @@ const BankAccounts = ({ navigation, route }) => {
   const [accountData, setAccountData] = useState()
   const [hasSelectedPaymentMethod, setHasSelectedPaymentMethod] = useState()
   const [modalVisible, setModalVisible] = useState(false);
-  const[activeBeneId, setActiveBeneID] = useState(null)
-
-
-  const gifUri = Image.resolveAssetSource(require('../../../assets/gif/loader.gif')).uri;
-
-
+  const [deleteAccountId, setDeleteAccountId] = useState('')
   const ternaryThemeColor = useSelector(
     state => state.apptheme.ternaryThemeColor,
   )
@@ -57,21 +50,7 @@ const BankAccounts = ({ navigation, route }) => {
     error: listAccountError,
     isLoading: listAccountIsLoading,
     isError: listAccountIsError
-  }] = useListAccountsMutation();
-
-
-  const [
-    updateStatusBankAccount,
-    {
-      data: updateStatusBankAccountData,
-      isLoading: updateStatusBankAccountIsLoading,
-      isError: updateStatusBankAccountIsError,
-      error: updateStatusBankAccountError,
-    },
-  ] = useUpdateStatusBankAccountMutation();
-
-
-  
+  }] = useListAccountsMutation()
 
   const [deleteBankFunc, {
     data: deleteBankData,
@@ -82,20 +61,27 @@ const BankAccounts = ({ navigation, route }) => {
 
   const focused = useIsFocused()
   const height = Dimensions.get('window').height
-  const deleteData = async (data) => {
+
+  const deleteData = async () => {
+    console.log("deleteData ",deleteAccountId)
     const credentials = await Keychain.getGenericPassword();
     if (credentials) {
       console.log(
         'Credentials successfully loaded for user ' + credentials.username
       );
       const token = credentials.username
-      const id = data
+      const id = deleteAccountId
 
       const params = {
         token: token,
-        id: id
+        id: Number(id)
       }
-      deleteBankFunc(params)
+      console.log(params)
+      if(deleteAccountId!=='')
+      {
+        deleteBankFunc(params)
+
+      }
       refetchData()
     }
   }
@@ -106,7 +92,6 @@ const BankAccounts = ({ navigation, route }) => {
         'Credentials successfully loaded for user ' + credentials.username
       );
       const token = credentials.username
-        
       const userId = userData.id
 
       const params = {
@@ -154,24 +139,9 @@ const BankAccounts = ({ navigation, route }) => {
   }, [deleteBankData, deleteBankError])
 
   useEffect(() => {
-    if (updateStatusBankAccountData) {
-      console.log("updateStatusBankAccountData", updateStatusBankAccountData)
-    
-      refetchData()
-    }
-    else if (updateStatusBankAccountError) {
-      console.log("updateStatusBankAccountError", updateStatusBankAccountError)
-    }
-  }, [updateStatusBankAccountData, updateStatusBankAccountError])
-
-
-  
-
-  useEffect(() => {
     if (listAccountData) {
       console.log("listAccountData", listAccountData.body)
       setAccountData(listAccountData.body)
-      
     }
     else if (listAccountError) {
       console.log("listAccountError", listAccountError)
@@ -186,6 +156,48 @@ const BankAccounts = ({ navigation, route }) => {
     setHasSelectedPaymentMethod(data.id)
   }
 
+  const DeleteAccountModal =()=>{
+    return(
+      <Modal
+          animationType="slide"
+          transparent={true}
+          visible={modalVisible}
+          onRequestClose={() => {
+
+            setModalVisible(!modalVisible);
+          }}>
+          <View style={styles.centeredView}>
+            <View style={styles.modalView}>
+              <Info name="info" color={ternaryThemeColor} size={40}></Info>
+              <PoppinsTextMedium style={{ color: 'black', width: 300, marginTop: 20, fontSize: 18 }} content="Are you sure?"></PoppinsTextMedium>
+              <PoppinsTextMedium style={{ color: 'black', width: '90%', marginTop: 20, fontSize: 20 }} content="Do you want to delete this account?"></PoppinsTextMedium>
+
+              <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: "center", marginTop: 20 }}>
+                <TouchableOpacity onPress={() => {
+                  console.log("cancel")
+                  setModalVisible(false)
+
+
+                }} style={{ alignItems: "center", justifyContent: "center", backgroundColor: 'white', flexDirection: "row", height: 40, width: 100, borderRadius: 20, borderWidth: 1 }}>
+                  <PoppinsTextMedium style={{ color: 'black', marginLeft: 10, fontWeight: '700' }} content="Cancel"></PoppinsTextMedium>
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => {
+                  // console.log("Upi component delete",props.id)
+                  console.log("Delete")
+                  setModalVisible(false)
+                  deleteData()
+
+                }} style={{ alignItems: "center", justifyContent: "center", backgroundColor: ternaryThemeColor, flexDirection: "row", marginLeft: 40, height: 40, width: 100, borderRadius: 20 }}>
+                  <PoppinsTextMedium style={{ color: 'white', marginLeft: 10 }} content="Delete"></PoppinsTextMedium>
+                </TouchableOpacity>
+              </View>
+
+            </View>
+          </View>
+        </Modal>
+    )
+  }
+
 
 
   const BankComponentAccount = (props) => {
@@ -196,11 +208,7 @@ const BankAccounts = ({ navigation, route }) => {
     const ifsc = props.ifsc
     const type = props.type
     const name = props.name
-  
-
     console.log("selected", props.unSelect)
-    
-
     useEffect(() => {
       console.log(props.unSelect)
       if (props.unSelect === "account") {
@@ -267,8 +275,8 @@ const BankAccounts = ({ navigation, route }) => {
                           <PoppinsTextMedium style={{color:ternaryThemeColor,fontSize:14,marginLeft:4}} content="Edit"></PoppinsTextMedium>                        
                         </TouchableOpacity> */}
               <TouchableOpacity onPress={() => {
-                console.log(props.id)
-                deleteData(String(props.id))
+                console.log("account component delete",props.id)
+                setDeleteAccountId(props.id)
                 setModalVisible(true)
 
               }} style={{ height: '50%', flexDirection: 'row', alignItems: "center", justifyContent: 'center' }}>
@@ -290,11 +298,8 @@ const BankAccounts = ({ navigation, route }) => {
   const BankComponentUpi = (props) => {
     const [selected, setSelected] = useState(false)
     const [openDrawer, setOpenDrawer] = useState(false)
-    const [isClicked,setIsClicked] = useState(false)
     const upi = props.upi
     const type = props.type
-    const status = props.status
-    const id = props.id
 
     useEffect(() => {
       console.log(props.unSelect)
@@ -303,30 +308,7 @@ const BankAccounts = ({ navigation, route }) => {
       }
     }, [props.unSelect])
 
-    useEffect(()=>{
-        console.log("isClicked", )
-    },[isClicked])
-
     const handleSelection = (selection) => {
-
-      const getToken = async()=>{
-        const credentials = await Keychain.getGenericPassword();
-        if (credentials) {
-          console.log(
-            'Credentials successfully loaded for user ' + credentials.username
-          );
-          const token = credentials.username
-
-          updateStatusBankAccount({
-            // s: getTenantData().tenant_id,
-            token: token,
-            id: id
-          })
-      }}
-      
-      getToken();
-     
-
       setSelected(!selection)
       console.log("selection",selection)
       if(!selection)
@@ -337,50 +319,16 @@ const BankAccounts = ({ navigation, route }) => {
           type: type
         })
       }
-      // setIsClicked(true)
+      
+      
 
-      
-      
+
+
     }
     return (
 
-      <View style={{ alignItems: "flex-start", justifyContent: 'center', width: '90%', height: 70, marginTop: 20, flexDirection: "row", borderBottomWidth: 1, borderColor: '#DDDDDD' }}>
-        <Modal
-          animationType="slide"
-          transparent={true}
-          visible={modalVisible}
-          onRequestClose={() => {
-
-            setModalVisible(!modalVisible);
-          }}>
-          <View style={styles.centeredView}>
-            <View style={styles.modalView}>
-              <Info name="info" color={ternaryThemeColor} size={40}></Info>
-              <PoppinsTextMedium style={{ color: 'black', width: 300, marginTop: 20, fontSize: 18 }} content="Are you sure?"></PoppinsTextMedium>
-              <PoppinsTextMedium style={{ color: 'black', width: '90%', marginTop: 20, fontSize: 20 }} content="Do you want to delete this account?"></PoppinsTextMedium>
-
-              <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: "center", marginTop: 20 }}>
-                <TouchableOpacity onPress={() => {
-                  console.log("cancel")
-                  setModalVisible(false)
-
-
-                }} style={{ alignItems: "center", justifyContent: "center", backgroundColor: 'white', flexDirection: "row", height: 40, width: 100, borderRadius: 20, borderWidth: 1 }}>
-                  <PoppinsTextMedium style={{ color: 'black', marginLeft: 10, fontWeight: '700' }} content="Cancel"></PoppinsTextMedium>
-                </TouchableOpacity>
-                <TouchableOpacity onPress={() => {
-                  console.log("Delete")
-                  setModalVisible(false)
-                  deleteData(String(props.id))
-
-                }} style={{ alignItems: "center", justifyContent: "center", backgroundColor: ternaryThemeColor, flexDirection: "row", marginLeft: 40, height: 40, width: 100,zIndex:1, borderRadius: 20 }}>
-                  <PoppinsTextMedium style={{ color: 'white', marginLeft: 10 }} content="Delete"></PoppinsTextMedium>
-                </TouchableOpacity>
-              </View>
-
-            </View>
-          </View>
-        </Modal>
+      <View style={{ alignItems: "flex-start", justifyContent: 'center', width: '90%', height: 110, marginTop: 20, flexDirection: "row", borderBottomWidth: 1, borderColor: '#DDDDDD' }}>
+        
         <View
           style={{
             width: '80%',
@@ -396,11 +344,7 @@ const BankAccounts = ({ navigation, route }) => {
             <TouchableOpacity onPress={() => {
               handleSelection(selected)
             }}>
-              {/* <Icon name={"checkcircle"} color={selected ? ternaryThemeColor : status && !isClicked ? ternaryThemeColor : "white"} size={20}></Icon> */}
-              {/* <Icon name={"checkcircle"} color={selected ? ternaryThemeColor : "white"} size={20}></Icon> */}
-              <Icon name={"checkcircle"} color={status ? ternaryThemeColor : "white"} size={20}></Icon>
-
-
+              <Icon name={"checkcircle"} color={selected ? ternaryThemeColor : "white"} size={20}></Icon>
             </TouchableOpacity>
           </View>
           {
@@ -411,7 +355,8 @@ const BankAccounts = ({ navigation, route }) => {
                          
                         </TouchableOpacity> */}
               <TouchableOpacity onPress={() => {
-                console.log(props.id)
+                console.log("ID",props.id)
+                setDeleteAccountId(props.id)
                 setModalVisible(true)
               }} style={{ height: '50%', flexDirection: 'row', alignItems: "center", justifyContent: 'center' }}>
                 <Delete style={{ marginLeft: 4 }} name="delete" size={14} color={ternaryThemeColor}></Delete>
@@ -428,7 +373,6 @@ const BankAccounts = ({ navigation, route }) => {
 
     );
   };
-
   return (
     <View
       style={{
@@ -438,7 +382,7 @@ const BankAccounts = ({ navigation, route }) => {
         backgroundColor: ternaryThemeColor,
         height: '100%',
       }}>
-
+        {modalVisible && <DeleteAccountModal></DeleteAccountModal>}
       {error && (
         <MessageModal
           modalClose={modalClose}
@@ -495,7 +439,7 @@ const BankAccounts = ({ navigation, route }) => {
           borderTopLeftRadius: 40,
           backgroundColor: '#F6F6F6',
           width: '100%',
-          height: height - 100,
+          height: '90%',
           marginTop: 30
         }}>
         <PoppinsTextMedium
@@ -507,7 +451,7 @@ const BankAccounts = ({ navigation, route }) => {
 
           }}
           content="Preferred Method"></PoppinsTextMedium>
-        <ScrollView style={{ width: '100%' }}>
+        <ScrollView contentContainerStyle={{width: '100%',height:'80%'}} style={{ width: '100%',height:'70%' }}>
           <View style={{ alignItems: "center", justifyContent: "center" }}>
 
             {listAccountData?.body?.length !== 0 &&
@@ -517,11 +461,11 @@ const BankAccounts = ({ navigation, route }) => {
                   justifyContent: 'flex-start',
                   width: '90%',
                   backgroundColor: 'white',
-                  borderColor: ternaryThemeColor,
-                  borderWidth: 2,
+                  borderColor: '#DDDDDD',
+                  borderWidth: 0.8,
                   borderRadius: 10,
                   marginTop: 20,
-                  // minHeight: 140
+                  minHeight: 140
                 }}>
                 {console.log("listAccountData in view", listAccountData?.body.length)}
                 {
@@ -529,7 +473,7 @@ const BankAccounts = ({ navigation, route }) => {
                     if (!Object.keys(item.bene_details).includes("upi_id")) {
                       console.log("true", item)
                       return (
-                        <BankComponentAccount unSelect={hasSelectedPaymentMethod} handleData={setSelectedPaymentMethod} key={index} type="account" id={item.id} bankName={item.bene_bank} accountNo={item.bene_details.bank_account} ifsc={item.bene_details.ifsc} name={item.bene_name} status={item.status=="1"}></BankComponentAccount>
+                        <BankComponentAccount unSelect={hasSelectedPaymentMethod} handleData={setSelectedPaymentMethod} key={index} type="account" id={item.id} bankName={item.bene_bank} accountNo={item.bene_details.bank_account} ifsc={item.bene_details.ifsc} name={item.bene_name}></BankComponentAccount>
 
                       )
 
@@ -537,7 +481,7 @@ const BankAccounts = ({ navigation, route }) => {
                     else {
                       console.log("false", item)
                       return (
-                        <BankComponentUpi unSelect={hasSelectedPaymentMethod} handleData={setSelectedPaymentMethod} key={index} id={item.id} type="upi" upi={item.bene_details.upi_id} status={item.status == "1"} ></BankComponentUpi>
+                        <BankComponentUpi unSelect={hasSelectedPaymentMethod} handleData={setSelectedPaymentMethod} key={index} id={item.id} type="upi" upi={item.bene_details.upi_id} ></BankComponentUpi>
                       )
                     }
 
@@ -550,46 +494,41 @@ const BankAccounts = ({ navigation, route }) => {
             }
 
             {(listAccountData?.body?.length == 0  || listAccountData == undefined ) && <View style={{ alignItems: 'center', marginTop:"60%"}}>
-              <PoppinsTextMedium style ={{fontSize:16}}content="No Bank Account has been added yet !"></PoppinsTextMedium>
+              <PoppinsTextMedium style ={{fontSize:16,color:'black'}}content="No Bank Account has been added yet !"></PoppinsTextMedium>
             </View>}
-
-            
-        
 
 
           </View>
-
+          
         </ScrollView>
-        <View style={{ flexDirection: "row", alignItems: 'center', justifyContent: 'center', position: 'absolute', bottom: 10, right: 20 }}>
+        <View style={{ flexDirection: "row", alignItems: 'center', justifyContent: 'center', position: 'absolute', bottom: 14, right: 20 }}>
           <PoppinsText content="Add Account" style={{ color: ternaryThemeColor, fontSize: 16 }}></PoppinsText>
           <TouchableOpacity onPress={() => { navigation.navigate('AddBankAccountAndUpi') }} style={{ backgroundColor: '#DDDDDD', height: 60, width: 60, borderRadius: 30, alignItems: "center", justifyContent: 'center', marginLeft: 10 }}>
 
             <Plus name="pluscircle" size={50} color={ternaryThemeColor}></Plus>
           </TouchableOpacity>
         </View>
-           {
-           type==="Cashback" &&  <TouchableOpacity onPress={()=>{
+        {
+           type==="Cashback" && <TouchableOpacity onPress={()=>{
+            if(hasSelectedPaymentMethod)
+            {
             navigation.navigate("OtpVerification",{type:"Cashback",selectedAccount:hasSelectedPaymentMethod})
-            }} style={{width:130,alignItems:'center',justifyContent:'center',backgroundColor:ternaryThemeColor,padding:8, position: 'absolute', bottom: 100, left: 145 }}>
+            }
+            }} style={{width:100,alignItems:'center',justifyContent:'center',backgroundColor:ternaryThemeColor,padding:8, position: 'absolute', bottom: 14, left: 20 }}>
               <PoppinsText content ="Get OTP" style={{color:'white',fontSize:16}}></PoppinsText>
             </TouchableOpacity>
             }
+        {/* {Platform.OS=='android'  && <View style={{ flexDirection: "row", alignItems: 'center', justifyContent: 'center', position: 'absolute', bottom: 10, right: 20 }}>
+          <PoppinsText content="Add Account" style={{ color: ternaryThemeColor, fontSize: 16 }}></PoppinsText>
+          <TouchableOpacity onPress={() => { navigation.navigate('AddBankAccountAndUpi') }} style={{ backgroundColor: '#DDDDDD', height: 60, width: 60, borderRadius: 30, alignItems: "center", justifyContent: 'center', marginLeft: 10 }}>
 
-{
-              listAccountIsLoading && <View style={{ backgroundColor: 'white', width:'100%', height:'150%' } }>
-                    <FastImage
-                        style={{ width: 100, height: 100, alignSelf: 'center',marginTop:'50%' }}
-                        source={{
-                            uri: gifUri, // Update the path to your GIF
-                            priority: FastImage.priority.normal,
-                        }}
-                        resizeMode={FastImage.resizeMode.contain}
-                    />
-                </View>
-            }
+            <Plus name="pluscircle" size={50} color={ternaryThemeColor}></Plus>
+          </TouchableOpacity>
+        </View>} */}
+          
+
 
       </View>
-      
 
     </View>
   );
